[refactor(crkbd)] moving code clutter out of keymap.c
